ravidosacorner/project/requirements.txt
asgiref==3.3.4
certifi==2021.5.30
dj-database-url==0.5.0
Django==3.2.4
django-heroku==0.3.1
gunicorn==20.1.0
Pillow==8.3.0
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
wincertstore==0.2
